Make commit information optional
