+++ packages/alfa-test-utils/src/tsconfig.json
@@ -11,9 +11,10 @@
     "audit/performance.ts",
     "audit/rules.ts",
     "report/logging.ts",
+    "report/get-rule-title.ts",
     "report/git.ts",
     "report/index.ts",
     "report/sip.ts",
-    "report/get-rule-title.ts"
+    "report/commit-information.ts"
   ]
 }
